GL: Initial texture class implementation
First go at a c++ style texture wrapper class for OpenGL. This
makes it nice and easy to abstract different textures that we
may need in the future.

Also note that this is a base class only. Another class must inherit
from this (so we can maintain tight type checking etc).